+++ javascript/webdriver/webdriver.js
@@ -36,6 +36,7 @@ goog.require('webdriver.Key');
 goog.require('webdriver.Locator');
 goog.require('webdriver.Serializable');
 goog.require('webdriver.Session');
+goog.require('webdriver.TouchSequence');
 goog.require('webdriver.logging');
 goog.require('webdriver.promise');
 goog.require('webdriver.until');
@@ -465,6 +466,23 @@ webdriver.WebDriver.prototype.quit = function() {
  */
 webdriver.WebDriver.prototype.actions = function() {
   return new webdriver.ActionSequence(this);
+};
+
+
+/**
+ * Creates a new touch sequence using this driver. The sequence will not be
+ * scheduled for execution until {@link webdriver.TouchSequence#perform} is
+ * called. Example:
+ *
+ *     driver.touchActions().
+ *         tap(element1).
+ *         doubleTap(element2).
+ *         perform();
+ *
+ * @return {!webdriver.TouchSequence} A new touch sequence for this instance.
+ */
+webdriver.WebDriver.prototype.touchActions = function() {
+  return new webdriver.TouchSequence(this);
 };
 
 
@@ -1758,6 +1776,19 @@ webdriver.WebElement.prototype.getDriver = function() {
  */
 webdriver.WebElement.prototype.getId = function() {
   return this.id_;
+};
+
+
+/**
+ * Returns the raw ID string ID for this element.
+ * @return {!webdriver.promise.Promise<string>} A promise that resolves to this
+ *     element's raw ID as a string value.
+ * @package
+ */
+webdriver.WebElement.prototype.getRawId = function() {
+  return this.id_.then(function(value) {
+    return value['ELEMENT'];
+  });
 };